Improve use-after-free detection for signature
